refactor(frontend/auth): 🧑‍💻 better imports
central export of exposed auth APIs
