+++ app/frontend/src/features/auth/index.ts
@@ -1,0 +1,5 @@
+import { AuthProvider } from "./context/auth-provider";
+import { useUser } from "./hooks/use-user";
+
+// * Export all auth related features
+export { AuthProvider, useUser };
